v0.7 update (NVSS) (#46)
* NVSS dualmode

Adds NVSS and fully tested with dualmode. Minor commits still to make

* Working dualmode and single mode version

A working version of adding NVSS and lots of other improvements. Likely one more commit will be needed for bug fixes

* Final v0.7 commit

Final commit to working v0.7. Partly to be able to start work on v0.8 that will streamline a lot of v0.7. Works, but messy.