Version 0.8 Transients Ratio Update (#62)
* First commit of transient ratio

This commit acts as a save point. Transient ratio works but no database integration as of yet

* Working transients ratio

Commit of a working version using the ratio of the sources fluxes as a transient identifier. Still bug fixes and clean ups to come

* Large and working update for v0.8

v0.8 working well with this commit, with overhauled website to better reflect the transient changes.

* Convolved flux error bug fix

* Main v0.8 update

Commit concentrates on enhancements to the website with added features such as: hotkeys, auto-next image using hotkeys, comments on images and candidates, galactic coordinates, locking of images unless claimed, csv file available on homepage. Also fixes RA wrap bug in plotting in the pipeline

* added SAMP coord broadcast to crossmatch detail (#60)

Accepting pull request to add SAMP functionality.

* Readme v0.8 update

* Settings template v0.8 update