User Query Feature (#68)
* Commit of working user query

The first commit of the user query which is in a basic working state. Next steps are to be able to browse through the results. Also updated comment fields to bootstrap4 and crispy forms.

* Working cone search

Added a cone search to the query form

* Queries now working with next and previous

Next and previous buttons are now working with user queries. Also added a possible associations table to highlight sources found in multiple images

* Added total number of results to query page

* Fixed comments redirect

* Final query commit

Added Source ID, RA and Dec fields along with a Quickview option on searches.

* Updated README